lrdp: rbtree: Support error recovery during tree iteration
In some cases it may be possible to detect whether an rbtree is corrupted.
Skip over any corrupted nodes, and their children, while iterating.
Additionally, log a message whenever a bad entry is detected. This behavior
may be enabled by the client by setting the 'logger' and 'debug' properties
for a particular instance of an RbTree.

Modify the --print-ionbuffer parser to use the new RbTree implementation.

Change-Id: I2ed2364b5e7d42ee009a0a49876feb82ab0fa78c