lrdp-v2: Refine MSM clock availability checking
LRDP-v2 is currently using the kernel version to judge if MSM
common clock is available on a system, which is inaccurate on
MSM8996_AU platform.

Change the clock dumping code to use the availability of macro
CONFIG_COMMON_CLK_MSM to judge if MSM common clock is used on
a platform.

Change-Id: Ic12ef5a249fb8b7c0f9633a6c8b9cdb23d0d39b4