lrdp-v2: Gdbmi interface clean up
When LRDP-v2 starts to parse RAM dumps, it will initialize pipes
to gdb instance, but on exit, the same pipes are not closed due
to __del__ is not invoked in ramdump.py. This will cause the
python environment raising the following exception, which will
not be handled by LRDP-v2 after it's unloaded:

close failed in file object destructor:
sys.excepthook is missing
lost sys.stderr

Add explicit gdbmi closure when RamDump instance is not used any
more, to avoid the unhandled exception.

Change-Id: I7cebd1418440ba56f9d7c4374286a5b088d948dc