lrdp: debug_image: Support guest VM cpu register dump tables
The hypervisor may populate an imem cookie with the location of
a standard dump table, instead of using the table allocated by linux.
Read the table location from this cookie instead of relying on the
linux variable 'memdump'.
On error, fall back to the old approach of reading from 'memdump'.

Change-Id: Id99bb643fb7f35de79ff7308a999cc2143d7aff0