ldrp_V2 : Support increased length in Linux banner string
This patch fixes the need for a hard coded value
to specify upper limit on length of the Linux Banner.

get_version() inside ramdump.py has been renamed to
get_matched_version() to reflect the true nature of
the function.

Added a variable banner_len in get_matched_version()
which is the length of linux_banner returned by
gdbmi.get_value_of_string('linux_banner').
This variable is further passed to read_cstring()
function to specify the max limit of characters to
be read. We are dynamically calculating the length
using gdb to read linux banner.

Change-Id: I571882a5999c26ccd8accd90b26d0149c839633b
Signed-off-by: Kumar Harsh <harkumar@codeaurora.org>