ldrp_V2 : Format event log data
This change formats the event log data. Timestamp is
converted to secs format. Since event log buffer is
circular in nature, added changes to sort the logs
in ascending timestamp order.

Any truncated log string is removed and not added
to the final output. Error handling is added in case
event zone is not part of persistent ram zone so
atleast console logs are extracted properly.

Change-Id: Id284df1646872b179e1cdfd3fe9ee68e7397c398
Signed-off-by: Kumar Harsh <harkumar@codeaurora.org>